Fix a deadlock in EglRenderer.releaseEglSurface.
Main thread is waiting for an operation on the render thread to complete
while holding the handler lock. Something can be waiting on the render
thread for this lock. This CL changes the behaviour so that the lock
is released before waiting for the operation to complete.

BUG=webrtc:6602,webrtc:6470
R=magjed@webrtc.org

Review-Url: https://codereview.webrtc.org/2449693003
Cr-Original-Commit-Position: refs/heads/master@{#14773}
Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: d0af5c6fd46d9248113a4c5f0e0fc54c9671b394